Everything is alittle everywhere, as i am currently reading op on adding C# operations to the Options component. Hence why everything is alittle wonky, as it's not done yet, and most things aren't even connected yet
Day was changed to time, and now only contains day and night. Even though there's eight options for time, daytime is the only one that changes the price. The rest use the same price.

Created two cs files for variables: BasePrice to get the prices to calculate the cost of the trip, and input which the form in Options is supposed to use.